% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_metrics.R
\name{calculateUC50}
\alias{calculateUC50}
\title{Calculate the UC50 (the number of unique clones which make up the top 50\% of the sample's
abundance) from a vector clonal abundances.}
\usage{
calculateUC50(abund)
}
\arguments{
\item{abund}{numeric vector of clonal abundances}
}
\value{
UC50
}
\description{
Calculate the UC50 (the number of unique clones which make up the top 50\% of the sample's
abundance) from a vector clonal abundances.
}
