% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_control_voc.R
\name{cellTypeControlVoc}
\alias{cellTypeControlVoc}
\title{standardizes spelling and fixes some common error in cell type strings. Returns all string that it
cannot match to a control vocabulary list as "Unk"}
\usage{
cellTypeControlVoc(...)
}
\arguments{
\item{...}{Any number of strings}
}
\value{
A factor with 6 level
}
\description{
standardizes spelling and fixes some common error in cell type strings. Returns all string that it
cannot match to a control vocabulary list as "Unk"
}
