% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_control_voc.R
\name{cellTypeControlVoc2}
\alias{cellTypeControlVoc2}
\title{standardizes spelling and fixes some common error in cell type strings. Returns all string that it
cannot match to a control vocabulary list unchanged.}
\usage{
cellTypeControlVoc2(...)
}
\arguments{
\item{...}{Any number of strings}
}
\value{
A factor with a variable number of level
}
\description{
standardizes spelling and fixes some common error in cell type strings. Returns all string that it
cannot match to a control vocabulary list unchanged.
}
