% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{get_doc_format}
\alias{get_doc_format}
\title{Get current document format.}
\usage{
get_doc_format()
}
\value{
The current document format, might be live (for Rstudio), latex or html.
}
\description{
Get current document format.
}
