% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_metrics.R
\name{get_pop_metrics}
\alias{get_pop_metrics}
\title{Calculate population metric for an intSite dataframe. Will combine samples with the same timePoint,
cellType and patient}
\usage{
get_pop_metrics(dd)
}
\arguments{
\item{dd}{An intSite dataframe}
}
\value{
A dataframe
}
\description{
Calculate population metric for an intSite dataframe. Will combine samples with the same timePoint,
cellType and patient
}
