% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_metrics.R
\name{get_pop_metrics_gtsp}
\alias{get_pop_metrics_gtsp}
\title{Calculate population metric for an intSite dataframe. Will NOT combine samples with the same
timePoint, cellType and patient.}
\usage{
get_pop_metrics_gtsp(dd, round = 3)
}
\arguments{
\item{dd}{An intSite dataframe}

\item{round}{Integer. the number of significant digits to show for maxRelAbund, Shannon, Gini,
and Simpson}
}
\value{
A dataframe
}
\description{
Calculate population metric for an intSite dataframe. Will NOT combine samples with the same
timePoint, cellType and patient.
}
