% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_table}
\alias{print_table}
\title{Print single page table}
\usage{
print_table(table, doc_format = get_doc_format())
}
\arguments{
\item{table}{a dataframe}

\item{doc_format}{print format, might be live (for Rstudio), latex or html}
}
\value{
strings to print the table
}
\description{
Print single page table
}
