% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table.R
\name{print_table_long}
\alias{print_table_long}
\title{print a multiple page table}
\usage{
print_table_long(
  table,
  doc_format = "live",
  font_size = 7,
  stripe_index = NULL,
  linesep = (if (.data$booktabs) c("", "", "", "", "\\\\addlinespace") else
    "\\\\hline")
)
}
\arguments{
\item{table}{a dataframe.}

\item{doc_format}{print format, might be live (for Rstudio), latex or html}

\item{font_size}{font size, the default is 7, might need a smaller font to fit long rows}

\item{stripe_index}{A numeric vector describing which rows should be shaded. default if every other row}

\item{linesep}{A string vector describing spacing between rows}
}
\value{
strings to print the table
}
\description{
print a multiple page table
}
